Refactored using AI's isValiDateTime method.
Checked with AI about whether each method should be static

Fixed some minor code issues at PHPStorm's prompting

Removed redundant setting of headers (JSONResponse already does that - verified with AI as I suspected it did)

Adds a first test for a custom window. Now need to add some validation for that.

Removes the empty results arrays; there are a number of server errors that are going to return without results - the app will need to account for that